backend/requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-docx
PyMuPDF
jinja2